Add signature image

+++ ios/RNTPTDocumentView.h
@@ -737,10 +737,8 @@ static NSString * const PTSignaturesManager_signatureDirectory = @"PTSignaturesM
 // Hygen Generated Methods
 - (void)setStampImageData:(NSString *)annotationId pageNumber:(NSInteger)pageNumber stampImageDataUrl:(NSString *)stampImageDataUrl;
 
-- (void)addAnnotation:(NSString *_Nonnull)type fieldName:(NSString *_Nonnull)fieldName pageNumber:(NSInteger)pageNumber x1:(double)x1 y1:(double)y1 x2:(double)x2 y2:(double)y2;
-
+- (void)addAnnotation:(NSString *_Nonnull)type fieldName:(NSString *_Nonnull)fieldName pageNumber:(int)pageNumber x1:(double)x1 y1:(double)y1 x2:(double)x2 y2:(double)y2;
 @end
-
 
 @interface RNTPTThumbnailsViewController : PTThumbnailsViewController
